Fix Roux's dumb mistakes
